refactor: extract inner classes and geometry methods from ManipulationHandle3D
Reduce ManipulationHandle3D.java from 664 to 488 lines by extracting:
- HandleGeometryHelper: 6 static geometry utility methods (package-private)
- Not3dHandleCriterion: named Criterion<Component> replacing anonymous class
- DoubleInterruptibleAnimation: top-level abstract class (public)
- Color4fInterruptibleAnimation: top-level abstract class (package-private)

Removed private methods (getScalable, invertParentScale, getObjectScale) and
replaced with static delegates. All 580 tests pass including 78 new extraction
tests. No subclass changes required — all in same package.

Closes #663
